+++ CHANGELOG.md
@@ -18,6 +18,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 - Update minor and patch NPM dependencies
 - Support `type` property on the buttons of the `<Dialog.Footer />` core component
+- Update text color on `ProposalVotingBodyContent` module component
 
 ## [1.0.57] - 2024-11-29